Fix PHP Insights for OAuth provider doubles
